# api/swagger/v1/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hospitalApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Department/{id}": {
      "get": {
        "tags": [
          "Department"
        ],
        "operationId": "Department_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentOutputDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Department"
        ],
        "operationId": "Department_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Department"
        ],
        "operationId": "Department_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DepartmentStaff": {
      "get": {
        "tags": [
          "DepartmentStaff"
        ],
        "operationId": "DepartmentStaff_GetAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DepartmentStaff"
        ],
        "operationId": "DepartmentStaff_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/DepartmentStaff/{id}": {
      "get": {
        "tags": [
          "DepartmentStaff"
        ],
        "operationId": "DepartmentStaff_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DepartmentStaff"
        ],
        "operationId": "DepartmentStaff_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.DepartmentStaffInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DepartmentStaff"
        ],
        "operationId": "DepartmentStaff_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Staff/{id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "hospitalApi.DTOs.Inputs.DepartmentInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "hospitalApi.DTOs.Inputs.DepartmentStaffInputDto": {
        "type": "object",
        "properties": {
          "fkStaffId": {
            "type": "integer",
            "format": "int32"
          },
          "fkDepartmentId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "hospitalApi.DTOs.Inputs.StaffInputDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "lastname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "fkRoleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.DepartmentOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "hospitalApi.DTOs.Outputs.DepartmentStaffOutputDto": {
        "required": [
          "department",
          "staff"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "department": {
            "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.DepartmentOutputDto"
          },
          "staff": {
            "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.StaffOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "lastname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "fkRoleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}